OPTICKS-294: Make the coding style consistent across all source code
git-svn-id: http://opticks.svn.sourceforge.net/svnroot/opticks/trunk/4.2.X@4773 6d0fddd1-9640-0410-91f9-aa98d5e64d34

+++ Code/application/Gui/ScriptingWidget.cpp
@@ -211,14 +211,14 @@ void ScriptingWidget::keyPressEvent(QKeyEvent* e)
             else if (e->key() == Qt::Key_Down)
             {
                // Let user go one past the end of the stack to clear command line.
-               if (mCommandIndex < (int) (mCommandStack.count()))
+               if (mCommandIndex < mCommandStack.count())
                {
                   mCommandIndex++;
                }
             }
 
             QString strCommand;
-            if (mCommandIndex >= 0 && mCommandIndex < (int) (mCommandStack.count()))
+            if (mCommandIndex >= 0 && mCommandIndex < mCommandStack.count())
             {
                strCommand = mCommandStack[mCommandIndex];
             }
